Add detection for unsupported Broadlink devices.
The code that sets up a connection to a Broadlink device, now detects when
the device type is reported as 'Unknown'. In such case, the library can talk
to the device, but the device itself is not (yet) recognized by the
python-broadlink module.

The warning that now is shown in such case, includes the device ID that will
have to be added to the broadlink.SUPPORTED_TYPES dict.

This change also includes improved logging for mqtt_transmit(). When the
internal call to device.transmit() raises an Exception, then the Exception
message is logged as a warning. This warning is what initially lead me to the
issue with unsupported devices.