Implement configurable Broadlink device types
Broadlink devices have a device type identifier, defining what type of
device it is. These identifiers are used by the python-broadlink module
to determine what kind of devices it is handling.

Sometimes, new identifiers are used for existing hardware, e.g. because
the manufacturer has changed. In such case, the python-broadlink module
has to be updated with the new identifiers to make things work.

This change allows users to define additional device types in the
broadlink-bridge config.ini file. This makes it possible to use new
device types, even before the python-broadlink module has been updated.